test(agent): cover execution through drawer UI
